Replace /var/www/app with /var/www/main
/var/www/app/app is kind of annoying for Symfony applications.

This change suggests that 'main' is your application; still a
generic name, because you should usually only have one application
in the container.

It was tempting to just make it /var/www, but that could get
tricky if there's a 1 in 100 case where you need /var/www/main
and /var/www/something-else.