Move from /var/www to /app
I'm conflicted a bit about this; is it better to use the Linux
filesystem conventions (e.g. FHS) or to make it "simple"?

/app is much less typing and easier to work with in Docker.
/var/www/app might be more correct (or just /var/www). But that
doesn't seem to be the Docker way.

For more complex containers it's possible to override where
/var/www/ (or anything else) makes more sense, but mostly it
should be one container, one app, spititng out one log stream.

So /app for now.